% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargarAccion.R
\name{descargarAccion}
\alias{descargarAccion}
\title{Descargar Accion}
\usage{
descargarAccion(nemo, fecha.ini = Sys.Date()-182, fecha.fin = Sys.Date(), nombre.archivo = nemo)
}
\arguments{
\item{nemo}{Una cadena de texto o vector de cadenas de texto con el nemotecnico de la accion a descargar.}

\item{fecha.ini}{Una fecha inicial para comenzar la descarga.}

\item{fecha.fin}{Una fecha final para la descarga.}

\item{nombre.archivo}{Establece un nombre para el archivo descargado. Por defecto, es el nemotecnico.}
}
\value{
Devuelve un objeto xts con los datos de la accion descargada.
}
\description{
Descarga una o varias acciones para un rango de fechas especificado.
}
\note{
Por defecto, si no se ingresan fechas para fecha.ini y fecha.fin, descarga los precios de los ultimos 182 dias.
}
\examples{
descargarAccion("TERPEL")
descargarAccion(c("TERPEL","EEB")))
descargarAccion("TERPEL",fecha.ini="2015-12-31",fecha.fin="2016-12-31")
}
\author{
John Fredy Arias
}
