% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargarAccion.R
\name{eliminarArchivosXLS}
\alias{eliminarArchivosXLS}
\title{eliminarArchivosXLS}
\usage{
eliminarArchivosXLS(ruta = paste0(getwd(), "/Datos"))
}
\arguments{
\item{ruta}{Establece la ruta donde se encuentran los archivos de excel a descargar.}
}
\description{
Una vez que se haya compilado la informacion en el archivo csv mediante
la funcion procesarArchivosXLS, se eliminan los arhivos de excel descargados de la BVC.
}
\examples{
eliminarArchivosXLS()
}
\author{
John Fredy Arias
}
