% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargarAccion.R
\name{procesarArchivosXLS}
\alias{procesarArchivosXLS}
\title{procesarArchivosXLS}
\usage{
procesarArchivosXLS(ruta = paste0(getwd(), "/Datos"), dec, sep)
}
\arguments{
\item{ruta}{Establece  la ruta para procesar los archivos de excel descargados de la BVC.}

\item{dec}{Establece el simbolo con el cual se deben separar los decimales en el archivo csv que se va a crear.}

\item{sep}{Establece el simbolo con el cual se deben separar las celdas en el archivo csv que se va a crear.}
}
\description{
Procesa todos los archivos de excel descargados de la pagina de la BVC, y compila toda
la informacion en un unico archivo csv.
}
\examples{
procesarArchivosXLS()
}
\author{
John Fredy Arias
}
