% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargarAccion.R
\name{procesarInformacion}
\alias{procesarInformacion}
\title{procesarInformacion}
\usage{
procesarInformacion(ruta = paste0(getwd(), "/Datos"))
}
\arguments{
\item{ruta}{Establece una ruta para buscar la informacion de los archivos.}
}
\description{
Consulta el tamanho de los archivos descargados.
}
\examples{
procesarInformacion()
}
\author{
John Fredy Arias
}
